refactor: simplify source_popular slice.
