--- lib/features/source_popular/domain/domain.dart
@@ -1,1 +1,0 @@
-export 'services/fetch_popular.dart';
