--- lib/features/source_popular/presentation/presentation.dart
@@ -1,7 +1,0 @@
-// Pages.
-export 'pages/popular_page.dart';
-
-// Providers.
-export 'providers/popular_fetch_family.dart';
-export 'providers/popular_family.dart';
-export 'providers/popular_view_family.dart';
